Allow launching the nodejs worker with inspector
To enable passing the `--inspector` flag to the nodejs worker, a support
was added to pass options to the worker executable.

After this commit, the nodejs inspector debugger is enabled when the
`AzureWebJobsEnv` is set to value `Development`.

This feature is useful when debugging the entire stack and allows stepping
through the code across programming language boundaries.